+++ container/datagen/requirements.txt
@@ -1,0 +1,4 @@
+black==22.8.0
+flake8==5.0.4
+mypy==0.971
+isort==5.10.1
