--- datagen/get_clicks_purchases.py
@@ -1,104 +1,0 @@
-from faker import Faker
-import random
-import json
-from time import sleep
-from datetime import datetime
-from confluent_kafka import Producer
-from uuid import uuid4
-
-fake = Faker()
-
-
-# Generate a random user agent string
-def random_user_agent():
-    return fake.user_agent()
-
-
-# Generate a random IP address
-def random_ip():
-    return fake.ipv4()
-
-
-# Generate a click event with the current datetime_occured
-def generate_click_event(user_id, product_id=None):
-    click_id = str(uuid4())
-    product_id = product_id or str(uuid4())
-    product = fake.word()
-    price = fake.pyfloat(left_digits=2, right_digits=2, positive=True)
-    url = fake.uri()
-    user_agent = random_user_agent()
-    ip_address = random_ip()
-    datetime_occured = datetime.now()
-
-    click_event = {
-        "click_id": click_id,
-        "user_id": user_id,
-        "product_id": product_id,
-        "product": product,
-        "price": price,
-        "url": url,
-        "user_agent": user_agent,
-        "ip_address": ip_address,
-        "datetime_occured": datetime_occured.strftime("%Y-%m-%d %H:%M:%S.%f")[
-            :-3
-        ],
-    }
-
-    return click_event
-
-
-# Generate a checkout event with the current datetime_occured
-def generate_checkout_event(user_id, product_id):
-    payment_method = fake.credit_card_provider()
-    total_amount = fake.pyfloat(left_digits=3, right_digits=2, positive=True)
-    shipping_address = fake.address()
-    billing_address = fake.address()
-    user_agent = random_user_agent()
-    ip_address = random_ip()
-    datetime_occured = datetime.now()
-
-    checkout_event = {
-        "checkout_id": str(uuid4()),
-        "user_id": user_id,
-        "product_id": product_id,
-        "payment_method": payment_method,
-        "total_amount": total_amount,
-        "shipping_address": shipping_address,
-        "billing_address": billing_address,
-        "user_agent": user_agent,
-        "ip_address": ip_address,
-        "datetime_occured": datetime_occured.strftime("%Y-%m-%d %H:%M:%S.%f")[
-            :-3
-        ],
-    }
-
-    return checkout_event
-
-
-# Function to push the events to a Kafka topic
-def push_to_kafka(event, topic):
-    producer = Producer({'bootstrap.servers': 'kafka:9092'})
-    producer.produce(topic, json.dumps(event).encode('utf-8'))
-    producer.flush()
-
-
-for _ in range(100000000):
-    user_id = random.randint(1, 100)
-    click_event = generate_click_event(user_id)
-    push_to_kafka(click_event, 'clicks')
-
-    # simulate multiple clicks & checkouts 50% of the time
-    while random.randint(1, 100) >= 50:
-        # sleep(1)
-        click_event = generate_click_event(user_id, click_event['product_id'])
-        push_to_kafka(click_event, 'clicks')
-
-        # sleep(1)
-        push_to_kafka(
-            generate_checkout_event(
-                click_event["user_id"], click_event["product_id"]
-            ),
-            'checkouts',
-        )
-
-    # sleep(1)
